Split up ComplianceBanner into modules
