+++ packages/react-components/src/components/ComplianceBanner/index.tsx
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 export { _ComplianceBanner } from './ComplianceBanner';
-export type { _ComplianceBannerProps, _ComplianceBannerStrings } from './ComplianceBanner';
+export type { _ComplianceBannerProps, _ComplianceBannerStrings } from './types';
